quick fix for last Iot test

+++ contracts/IotSimulation.sol
@@ -25,14 +25,14 @@ contract IotSimulation is Destructible{
      * @param id Vehicle identification number
      * @return result Execution result
      */
-    function generateIotOutput(uint24 id, uint salt) returns (bool result) {
+    function generateIotOutput(uint id, uint salt) returns (bool result) {
         require(id != 0);
         require(carAssetLogicAddr != address(0));
 
         uint number = id + salt;
         CarAssetLogicInterface carAssetLogic = CarAssetLogicInterface(carAssetLogicAddr);
         carAssetLogic.updateViaIotSimulator(
-            id,
+            uint24(id),
             generateLongitudeResult(number),
             generateLatitudeResult(number),
             "/link"
